Pretty much edited most of this so I could fit in our listviews inside of the separate ListFragments.
The service allows the device to play music while the app is closed or the display is turned off.

Added 3 buttons. (Unused, I think)

Added a textfor to show the album for each song, too.

Sorry for some code removal, but it got confusing. Runs, but haven't tested playing the music yet since my laptop isn't capable of that.