Created a Playlist class that extended ArrayList<Song>.
Converted all types of ArrayList<Song> to Playlist except for original
ArrayList<Song> of songs from getMusic();

Created method in MainActivity solely for adding default playlists/tabs
so in the future, can be edited easily if needed.

Added a method in MusicService to get the song's index from the
current playlist, and find the index of that song in the main song collection
so the song plays the correct song. (I can explain this if needed)

Some other changes